yanked POST /api/categories, moved it's logic into svcCreateCategory
